Update metricConverter.js
Dearest friend,
I allow myself to present a slight edit of your already clever solution. And object containing all possible actions make the if-else statements of no use. Also I moved the toFixed function in the end in order less code to be repeated. Feel free to discharge or edit my proposal on it's own.